+++ packages/flixlix-cards/energy-flow-card-plus/src/ui-editor/schema/_schema-all.ts
@@ -150,12 +150,12 @@ export const advancedOptionsSchema = memoizeOne((localize, displayZeroLinesMode:
       },
       {
         name: "w_decimals",
-        label: "Watt Decimals",
+        label: "Wh Decimals",
         selector: { number: { mode: "box", min: 0, max: 5, step: 1 } },
       },
       {
         name: "kw_decimals",
-        label: "kW Decimals",
+        label: "kWh Decimals",
         selector: { number: { mode: "box", min: 0, max: 5, step: 1 } },
       },
       {
@@ -169,18 +169,8 @@ export const advancedOptionsSchema = memoizeOne((localize, displayZeroLinesMode:
         selector: { number: { mode: "box", min: 0, max: 1000000, step: 0.01 } },
       },
       {
-        name: "max_expected_power",
-        label: "Max Expected Power (in Watts)",
-        selector: { number: { mode: "box", min: 0, max: 1000000, step: 0.01 } },
-      },
-      {
-        name: "min_expected_power",
-        label: "Min Expected Power (in Watts)",
-        selector: { number: { mode: "box", min: 0, max: 1000000, step: 0.01 } },
-      },
-      {
         name: "watt_threshold",
-        label: "Watt to Kilowatt Threshold",
+        label: "Wh to kWh Threshold",
         selector: { number: { mode: "box", min: 0, max: 1000000, step: 1 } },
       },
 
@@ -199,12 +189,6 @@ export const advancedOptionsSchema = memoizeOne((localize, displayZeroLinesMode:
         name: "no_labels",
         label: localize("editor.no_labels"),
         selector: { boolean: {} },
-      },
-      {
-        name: "use_new_flow_rate_model",
-        label: "New Flow Model?",
-        selector: { boolean: {} },
-        default: true,
       },
       {
         name: "sort_individual_devices",