+++ packages/flixlix-cards/energy-flow-card-plus/CHANGELOG.md
@@ -1,5 +1,17 @@
 # energy-flow-card-plus
 
+## 0.2.0
+
+### Minor Changes
+
+- rewrite with monorepo
+
+### Patch Changes
+
+- Updated dependencies [655d4e8]
+- Updated dependencies [ea6f3d6]
+  - @flixlix-cards/shared@0.0.1
+
 ## 0.3.2
 
 ### Patch Changes